ploop: copyFile(): initiate fdatasync() on copied file
When copying a ploop image (from a master one), it makes sense
to run fdatasync(2) on destination file contents, because
for ploop mount (ADD_DELTA) kernel wants the file to be
physically written to disk, so it causes the flushing anyway,
and by doing this proactively we can speed up the mounting.

Note we are using fdatasync(2) not fsync(2) because we are not
interested in file metadata being flushed to disk, we only need
file data, and fdatasync() can be faster and cause less I/O.

Note this might potentially lead to "too many open files" error
as destination file closing is done asyncronously.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>