ploop: fix a linker warning
When libxml2 is linked into a static docker binary, linker emits
the the following warning:

> ---> Making bundle: binary (in bundles/1.7.1/binary)
> Building: bundles/1.7.1/binary/docker-1.7.1
> # github.com/docker/docker/docker
> /usr/lib/gcc/x86_64-linux-gnu/4.8/../../../x86_64-linux-gnu/libxml2.a(nanohttp.o): In function `xmlNanoHTTPConnectHost':
> (.text+0x7ef): warning: Using 'getaddrinfo' in statically linked applications requires at runtime the shared libraries from the glibc version used for linking
> /usr/lib/gcc/x86_64-linux-gnu/4.8/../../../x86_64-linux-gnu/libxml2.a(nanohttp.o): In function `xmlNanoHTTPConnectHost':
> (.text+0x8a4): warning: Using 'gethostbyname' in statically linked applications requires at runtime the shared libraries from the glibc version used for linking
> Created binary: bundles/1.7.1/binary/docker-1.7.1

This warning is caused by the fact that GLIBC resolver depends on dynamic
libs (/lib/libnss_*) even in case glibc and the binary that uses it are
compiled statically -- and so the resulting binary might need those libs
on a system where it runs.

Still, the warning is harmless in this case as the only libxml2 user is
libploop and it does not use libxml2 functionality to connect to http hosts.

Nevertheless, it looks pretty scary, and the way to remove this warning
is to remove its cause -- i.e. recompile libxml2 without http client support.

This is what this patch does -- for the static case only.

NB: I am using libxml2-2.2.9 which is currently the latest version,
while Ubutu 14.04 comes with libxml2-2.2.8 plus some patches. I checked
those patches to make sure the relevant ones (such as CVE fixes) are
already included in 2.2.9, while non-included ones are about
the building process etc.

Signed-off-by: Kir Kolyshkin <kir@openvz.org>