ploop: decrease lock contention in Get()/Put()
We used to hold the lock for the whole duration of Get()/Put(),
not just to access mounts map. This lead to lock contention.

This commit fixes the problem, and the resulting binary survived
a few hours of stress test done by parallel run of

	crosbymichael/docker-stress --containers 100 -c 2 -k 5s
	spotify/docker-stress -c 50 -t 5
	make # docker build
	while true; do docker info; sleep 1; done

That doesn't prove anything of course, but it seems it's working.

Some performance figures using github.com/crosbymichael/docker-stress:

  before:
	./docker-stress --containers 10 -c 20 -k 5s
	INFO[0033] ran 22 containers in 33.567809 seconds (0.655390 per sec.)

  after:
	./docker-stress --containers 10 -c 20 -k 5s
	INFO[0012] ran 22 containers in 12.919807 seconds (1.702812 per sec.)

Signed-off-by: Kir Kolyshkin <kir@openvz.org>