integration-cli/TestLogsAPIFollowEmptyOutput: fix
The original meaning of this test case was to check that the HTTP
responce from the daemon comes before the container stops, even
in case the container has no output (see commit 589de35651ce8c91
and issue #12704 for details).

Now, commit 00f65ae810679d587 changes the timeout to wait for
the HTTP response to be greater than the container lifetime,
thus defeating the original test purpose.

Fix it, while keeping in mind that CI might be slow (and so
the timeouts should be generous enough).

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>